Añadida autenticacion en API REST Añade metodo para obtener lista de proyectos
