+++ openhrapi/main.py
@@ -1,70 +1,36 @@
 from random import random
 from time import sleep
 
-import requests
-from bs4 import BeautifulSoup
-from fake_useragent import UserAgent
+from flask import Flask
 
-from openhrapi.config import URL_FICHAJE
-
-from flask import Flask, request
+from openhrapi.navigate import get_logged_session, post_fichaje, get_proyectos
+from flask_httpauth import HTTPBasicAuth
 
 app = Flask(__name__)
+auth = HTTPBasicAuth()
+
+
+@auth.verify_password
+def verify_password(username, password):
+    try:
+        s = get_logged_session(usuario=username, password=password)
+    except ValueError as ex:
+        app.logger.info(ex)
+        app.logger.debug(username, password)
+        return False
+
+    sleep(random() * 5 + 5)
+    return s
 
 
 @app.route('/ficha', methods=['POST'])
+@auth.login_required
 def fichar():
-
-    user = request.values.get('user')
-    password = request.values.get('password')
-
-    try:
-        s = get_logged_session(usuario=user, password=password)
-    except ValueError as ex:
-        app.logger.info(ex)
-        app.logger.debug(user, password)
-        return 'Not authorized', 401
-
-    sleep(random()*5+5)
-
-    post_fichaje(session=s)
+    post_fichaje(session=auth.current_user())
     return {}
 
 
-def get_logged_session(usuario, password):
-    ua = UserAgent().random
-    session = requests.session()
-    session.headers['User-Agent'] = ua
-
-    response = session.get(URL_FICHAJE)
-    soup = BeautifulSoup(response.text, features="html.parser")
-    form = soup.form
-
-    campo = form.find_all('input')
-    data = {c['name']: c.get('value') for c in campo if 'name' in c.attrs}
-
-    data['FrmEntrada[usuario]'] = usuario
-    data['FrmEntrada[clave]'] = password
-
-    r = session.post(form['action'], data=data)
-
-    if r.url.endswith('/autenticar/inicio'):
-        raise ValueError('Usuario, password no válido')
-
-    return session
-
-
-def post_fichaje(session):
-    response = session.get(URL_FICHAJE)
-
-    soup = BeautifulSoup(response.text, features="html.parser")
-    form = soup.form
-
-    campos = form.find_all('input')
-    data = {c['name']: c.get('value') for c in campos if c['type'] != 'submit'}
-    data['aPic[localizacion][txt]'] = 'No se ha permitido el acceso a la posición del usuario.'
-
-    boton = soup.find(lambda x: x.get('value') == 'Grabar' and x.get('type') == 'submit')
-    data[boton['name']] = boton['value']
-
-    return session.post(form['action'], data=data)
+@app.route('/proyectos', methods=['GET'])
+@auth.login_required
+def proyectos():
+    return {'proyectos': get_proyectos(session=auth.current_user())}
