+++ requirements.txt
@@ -20,6 +20,7 @@ colorama==0.4.4; platform_system == 'Windows'
 fake-useragent==0.1.11
 
 flask==2.0.1
+flask_httpauth
 
 idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
